Split into interactive and non-interactive tests
