+++ tests/rexpect/signals/mod.rs
@@ -1,0 +1,2 @@
+mod interactive;
+mod noninteractive;
